Inline getAppConfig into rootIndexPlugin
