+++ cli/src/plugins/rootIndexPlugin.js
@@ -1,5 +1,45 @@
-import { getAppConfig } from "@preact/benchmark-apps";
 import { getDepConfig } from "@preact/benchmark-deps";
+import { readdir } from "node:fs/promises";
+import { repoRoot } from "../utils.js";
+
+/** @type {() => Promise<RootConfig["apps"]>} */
+async function getAppConfig() {
+	const appNames = (await readdir(repoRoot("apps"), { withFileTypes: true }))
+		.filter(
+			(e) =>
+				e.isDirectory() && !e.name.startsWith("_") && e.name !== "node_modules"
+		)
+		.map((e) => e.name);
+
+	/** @type {RootConfig["apps"]} */
+	const apps = {};
+	for (let appName of appNames) {
+		const appContents = await readdir(repoRoot(`apps/${appName}/`), {
+			withFileTypes: true,
+		});
+
+		/** @type {Record<string, string>} */
+		const benchmarks = {};
+		/** @type {Record<string, string>} */
+		const implementations = {};
+		for (let dirEntry of appContents) {
+			if (dirEntry.isFile() && dirEntry.name.endsWith(".html")) {
+				benchmarks[dirEntry.name] = `apps/${appName}/${dirEntry.name}`;
+			}
+
+			if (dirEntry.isDirectory() && !dirEntry.name.startsWith("_")) {
+				implementations[dirEntry.name] = `apps/${appName}/${dirEntry.name}`;
+			}
+		}
+
+		apps[appName] = {
+			benchmarks,
+			implementations,
+		};
+	}
+
+	return apps;
+}
 
 /** @type {() => import('vite').Plugin} */
 export function rootIndexPlugin() {
